🎨 display balthazar sponsors in footer layout + all on specific sponsors page
