+++ pages/Sponsors.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { useHead, useNuxtApp, createError, useAPI } from "#imports";
+import { useHead, useNuxtApp, createError, useAPI, definePageMeta } from "#imports";
 import { SectionSponsorsBecomeSponsor, SectionSponsorsOffers } from "#components";
 import { ASSOCIATION_NAME, EDITION } from "~/services/constants";
 import type { Ref } from "vue";
@@ -11,6 +11,7 @@ if (!$SHOW_PAGE_SPONSORS) {
   throw createError({ statusCode: 404 });
 }
 
+definePageMeta({ layout: "page-without-sponsors" });
 useHead({ title: `Sponsors | ${ASSOCIATION_NAME}` });
 
 const { data }: { data: Ref<Offer[]> } = await useAPI("/offers", { params: {
@@ -26,5 +27,6 @@ data.value.sort((a, b) => parseFloat(a.price) - parseFloat(b.price));
       v-if="data.length !== 0"
       :offers="data"
     />
+    <SectionSponsors />
   </main>
 </template>
